+++ BE_TKDecor/Core/Dtos/ProductReport/ProductReportCreateDto.cs
@@ -4,7 +4,7 @@ namespace BE_TKDecor.Core.Dtos.ProductReport
 {
     public class ProductReportCreateDto
     {
-        public int ProductReportedId { get; set; }
+        public long ProductReportedId { get; set; }
 
         public string Reason { get; set; } = null!;
     }
