+++ BE_TKDecor/Core/Dtos/UserAddress/UserAddressUpdateDto.cs
@@ -4,7 +4,7 @@ namespace BE_TKDecor.Core.Dtos.UserAddress
 {
     public class UserAddressUpdateDto
     {
-        public int UserAddressId { get; set; }
+        public long UserAddressId { get; set; }
 
         [MaxLength(255)]
         public string FullName { get; set; } = null!;
